feat(android): multi-target FEBuilderGBA.Avalonia so the Android head builds an APK (#1121)
Make the Android head (FEBuilderGBA.Android) build a real APK against the shared
FEBuilderGBA.Avalonia UI by conditionally multi-targeting net9.0;net9.0-android
(Approach A, accepted + Copilot-reviewed).

User request: implement #1121 — make the Android head BUILD against the shared
FEBuilderGBA.Avalonia UI by multi-targeting net9.0;net9.0-android, opt-in via
EnableAndroidTarget (default OFF so desktop .sln + CI see net9.0 only), with the
Android head ProjectReference using AdditionalProperties="EnableAndroidTarget=true".

Changes:
- FEBuilderGBA.Avalonia.csproj: opt-in EnableAndroidTarget (default false) selects
  net9.0 vs net9.0;net9.0-android. Per-TFM OutputType (WinExe desktop / Library
  android), desktop-only ApplicationManifest, ANDROID DefineConstant on android,
  Avalonia.Desktop + Microsoft.CodeAnalysis.CSharp desktop-only, and Program.cs +
  GapSweep/** + App.GapSweep.cs excluded on android. config/** Content kept (loose
  on android, harmless; APK asset bundling deferred to #1123).
- App.GapSweep.cs (new): desktop-only partial holding every GapSweep-type-dependent
  method (RunGapSweep[Standalone], RunDensity/Labels/Jumps/Undo/L10n/Gallery/All
  sweeps + shared variants, BuildDefaultSubSweeps, SubSweepsOverride, SafePngCount,
  ComputeGalleryStatus, FindRepoRoot). Excluded on the android TFM.
- App.axaml.cs: removed the GapSweep using + the moved methods; kept the plain
  string?/bool gap-sweep static properties (assigned by ParseArgs); guarded the
  OnFrameworkInitializationCompleted RunGapSweep() dispatch behind #if !ANDROID;
  changed the dangling L10nScanner.DefaultLanguages cref to plain text.
- FEBuilderGBA.Android.csproj: ProjectReference now carries
  AdditionalProperties="EnableAndroidTarget=true"; comment updated (blocker resolved,
  still not in .sln). Build requires -p:EnableAndroidTarget=true because NuGet
  restore ignores per-reference AdditionalProperties (NETSDK1005 otherwise).
- Tests: AndroidMultiTargetDesktopParityTests asserts the desktop build is
  unchanged (gap-sweep statics still on App; RunGapSweepStandalone null for
  non-gap-sweep args; gap-sweep flag wiring intact).
- Docs: docs/ANDROID.md (§1 table, §7 build status + honest conclusion, §8/§9
  follow-ups), FEBuilderGBA.Android/README.md, docs/CROSS_PLATFORM.md updated —
  APK builds now; explicitly NOT device-validated (runtime port = #1122/#1123/#1124).

Validation (build-only; no Android device/emulator — explicit user decision):
- Desktop unchanged: Core/SkiaSharp/CLI build 0 errors; Avalonia builds net9.0
  ONLY (no net9.0-android dir → EnableAndroidTarget defaults OFF); full sln x86
  MSBuild green (0 errors); Avalonia.Tests 7983 pass (incl. 4 new); Core.Tests
  10 pre-existing FE8U skill-anime template failures only (no new); FEBuilderGBA.Tests
  1856 pass.
- Android: dotnet build FEBuilderGBA.Android.csproj -c Release -p:EnableAndroidTarget=true
  → Build succeeded, 0 errors; produced com.laqieer.febuildergba(-Signed).apk
  (~33 MB) with FEBuilderGBA.Avalonia.dll compiled under net9.0-android.

Implements #1121 — Android: multi-target FEBuilderGBA.Avalonia so the Android head builds an APK against the shared UI.

Approach A: conditional multi-target net9.0;net9.0-android, opt-in via EnableAndroidTarget
(default OFF so desktop .sln + CI see net9.0 only). Android head ProjectReference uses
AdditionalProperties="EnableAndroidTarget=true" to activate the android TFM on the shared project.
GapSweep (Roslyn dev-tooling) + Program.cs + Avalonia.Desktop + app.manifest are desktop-only.
Validation: desktop/build-only (no Android device/emulator — explicit user decision).

Tool: Claude Code (claude-opus-4-8[1m], Opus 4.8 1M context)
