+++ FEBuilderGBA.Android/README.md
@@ -1,8 +1,12 @@
-# FEBuilderGBA.Android (Avalonia.Android head skeleton)
+# FEBuilderGBA.Android (Avalonia.Android head)
 
-> **Exploration skeleton — not shipped, not runnable yet.** Part of epic
-> [#1070](https://github.com/laqieer/FEBuilderGBA/issues/1070). Read the full
-> feasibility assessment first: **[docs/ANDROID.md](../docs/ANDROID.md)**.
+> **Builds an APK; not yet runnable / device-validated.** Part of epic
+> [#1070](https://github.com/laqieer/FEBuilderGBA/issues/1070). The structural
+> prerequisite landed in
+> [#1121](https://github.com/laqieer/FEBuilderGBA/issues/1121) — this head now
+> builds a real APK against the shared, conditionally multi-targeted
+> `FEBuilderGBA.Avalonia`. Read the full feasibility assessment first:
+> **[docs/ANDROID.md](../docs/ANDROID.md)**.
 
 This is a minimal Avalonia.Android *head* that reuses the shared
 `FEBuilderGBA.Avalonia` UI (which transitively pulls in `FEBuilderGBA.Core` +
@@ -21,21 +25,36 @@ standalone.
 
 ```bash
 dotnet workload install android          # one-time
-dotnet build FEBuilderGBA.Android/FEBuilderGBA.Android.csproj
+dotnet build FEBuilderGBA.Android/FEBuilderGBA.Android.csproj -c Release -p:EnableAndroidTarget=true
 ```
+
+The `-p:EnableAndroidTarget=true` flag is **required**. `AdditionalProperties` on
+the `ProjectReference` correctly activates the android TFM for the *build* phase,
+but NuGet *restore* uses a separate static graph that ignores per-reference
+`AdditionalProperties`, so without the global property restore writes a
+net9.0-only assets file for `FEBuilderGBA.Avalonia` and the build fails with
+`NETSDK1005`. Passing it as a global property flows it to the referenced
+project's restore + build. A successful build emits
+`com.laqieer.febuildergba-Signed.apk` under
+`bin/Release/net9.0-android/`.
 
 ## Current build status (honest)
 
 - ✅ A minimal `net9.0-android` project builds + packages a signed APK with the
   workload installed (toolchain verified).
 - ✅ This head's own code (`MainActivity`) **compiles** against the shared
-  `FEBuilderGBA.Avalonia` reference; MSBuild builds `FEBuilderGBA.Avalonia` as
-  its own `net9.0` TFM (it is **not** recompiled under the Android TFM) and the
-  build reaches the Android RID-packaging stage.
-- ❌ The **full APK** does NOT build yet: `FEBuilderGBA.Avalonia` targets plain
-  `net9.0`, so per-RID (`android-arm64`/`android-x64`) packaging fails
-  (`NETSDK1047` / `NETSDK1112`). The fix is to multi-target / split the shared
-  UI — tracked as a follow-up under #1070. See
+  `FEBuilderGBA.Avalonia` reference.
+- ✅ **The full APK builds (#1121).** `FEBuilderGBA.Avalonia` conditionally
+  multi-targets `net9.0;net9.0-android` (opt-in via `EnableAndroidTarget`,
+  default OFF); the `ProjectReference` below activates the android TFM with
+  `AdditionalProperties="EnableAndroidTarget=true"` (an MSBuild-verified
+  requirement). `dotnet build … -c Release` produces an APK under
+  `bin/Release/net9.0-android/`. On the android TFM the shared project excludes
+  `Program.cs`, `Avalonia.Desktop`, `app.manifest`, the `GapSweep` Roslyn
+  dev-tooling, and `Microsoft.CodeAnalysis.CSharp`.
+- ⚠️ **Not yet device/emulator-validated** — it builds, but the runtime port
+  (single-view `MainView`, config asset extraction, SAF ROM I/O) is still
+  pending. See
   [docs/ANDROID.md §7](../docs/ANDROID.md#7-build-status-in-this-environment).
 
 ## Known skeleton limitations
@@ -47,4 +66,5 @@ dotnet build FEBuilderGBA.Android/FEBuilderGBA.Android.csproj
   multi-window model to page navigation is the largest follow-up.
 - `config/` asset bundling + first-run extraction and SAF-stream ROM I/O are
   separate follow-ups (the `<AndroidAsset>` line in the csproj is intentionally
-  commented out).
+  commented out — config bundling is deferred to #1123). Today `config/**` ships
+  as loose Content beside the APK, which is harmless but not the shipping layout.
