+++ FEBuilderGBA.Avalonia/FEBuilderGBA.Avalonia.csproj
@@ -1,26 +1,64 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <OutputType>WinExe</OutputType>
+    <!-- Android TFM is OPT-IN (#1121). Only the FEBuilderGBA.Android head activates it
+         (via ProjectReference AdditionalProperties="EnableAndroidTarget=true"). Default OFF
+         so the desktop .sln build + all CI (no android workload) see net9.0 only. -->
+    <EnableAndroidTarget Condition="'$(EnableAndroidTarget)' == ''">false</EnableAndroidTarget>
+  </PropertyGroup>
+  <PropertyGroup Condition="'$(EnableAndroidTarget)' == 'true'">
+    <TargetFrameworks>net9.0;net9.0-android</TargetFrameworks>
+  </PropertyGroup>
+  <PropertyGroup Condition="'$(EnableAndroidTarget)' != 'true'">
     <TargetFramework>net9.0</TargetFramework>
+  </PropertyGroup>
+  <PropertyGroup>
+    <!-- Desktop ships an executable; the android head consumes this project as a
+         library (it supplies its own MainActivity entry point), so OutputType is
+         per-TFM. -->
+    <OutputType Condition="'$(TargetFramework)' != 'net9.0-android'">WinExe</OutputType>
+    <OutputType Condition="'$(TargetFramework)' == 'net9.0-android'">Library</OutputType>
     <RootNamespace>FEBuilderGBA.Avalonia</RootNamespace>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     <BuiltInComInteropSupport>true</BuiltInComInteropSupport>
-    <ApplicationManifest>app.manifest</ApplicationManifest>
+    <!-- app.manifest is a Win32 application manifest — desktop-only. -->
+    <ApplicationManifest Condition="'$(TargetFramework)' != 'net9.0-android'">app.manifest</ApplicationManifest>
+    <!-- Defensive ANDROID compile constant for the #if !ANDROID guards (e.g. the
+         GapSweep dispatch in App.axaml.cs). The .NET Android SDK already defines
+         ANDROID, but we set it explicitly on the android TFM so the contract is
+         visible here; the desktop net9.0 TFM deliberately does NOT define it. -->
+    <DefineConstants Condition="'$(TargetFramework)' == 'net9.0-android'">$(DefineConstants);ANDROID</DefineConstants>
   </PropertyGroup>
   <ItemGroup>
     <InternalsVisibleTo Include="FEBuilderGBA.Avalonia.Tests" />
   </ItemGroup>
   <ItemGroup>
+    <!-- Avalonia + Fluent theme are cross-TFM (desktop + android). -->
     <PackageReference Include="Avalonia" Version="11.2.3" />
-    <PackageReference Include="Avalonia.Desktop" Version="11.2.3" />
     <PackageReference Include="Avalonia.Themes.Fluent" Version="11.2.3" />
-    <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.8.0" />
+    <!-- Avalonia.Desktop (classic desktop lifetime / StartWithClassicDesktopLifetime)
+         and Microsoft.CodeAnalysis.CSharp (Roslyn, used only by the desktop-only
+         GapSweep dev-tooling) are excluded on the android TFM (#1121). -->
+    <PackageReference Include="Avalonia.Desktop" Version="11.2.3" Condition="'$(TargetFramework)' != 'net9.0-android'" />
+    <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.8.0" Condition="'$(TargetFramework)' != 'net9.0-android'" />
+  </ItemGroup>
+  <ItemGroup Condition="'$(TargetFramework)' == 'net9.0-android'">
+    <!-- Desktop-only sources excluded on android (#1121):
+         - Program.cs           : [STAThread] desktop entry point (StartWithClassicDesktopLifetime).
+         - GapSweep\**          : Roslyn static-analysis dev-tooling (not shippable Android UI).
+         - App.GapSweep.cs      : the App partial holding the GapSweep-type-dependent methods. -->
+    <Compile Remove="Program.cs" />
+    <Compile Remove="GapSweep\**\*.cs" />
+    <Compile Remove="App.GapSweep.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\FEBuilderGBA.Core\FEBuilderGBA.Core.csproj" />
     <ProjectReference Include="..\FEBuilderGBA.SkiaSharp\FEBuilderGBA.SkiaSharp.csproj" />
   </ItemGroup>
-  <!-- Copy config/ directory to output, excluding patch2 (installed separately via git) -->
+  <!-- Copy config/ directory to output, excluding patch2 (installed separately via git).
+       On net9.0 (desktop) this lands beside the exe and Core reads it via
+       CoreState.BaseDirectory. On net9.0-android the loose Content is harmless but is
+       NOT the shipping mechanism — APK config bundling (AndroidAsset + first-run
+       extraction to Context.FilesDir) is deferred to #1123. -->
   <ItemGroup>
     <Content Include="..\config\**\*.*" Exclude="..\config\patch2\**" LinkBase="config" CopyToOutputDirectory="PreserveNewest" />
   </ItemGroup>
