+++ docs/CROSS_PLATFORM.md
@@ -199,4 +199,13 @@ msbuild /m /p:Configuration=Release /p:Platform=x86 /t:build /restore FEBuilderG
 
 Running the Windows binary under emulation gives you the *desktop* program on an Android device; it does **not** give you a touch-native app. An actual native Android build of the Avalonia GUI is a substantial, separate port tracked as the exploration epic [#1070](https://github.com/laqieer/FEBuilderGBA/issues/1070) — there is no commitment to ship it yet.
 
-The full, evidence-backed feasibility assessment lives in **[docs/ANDROID.md](ANDROID.md)**: which layers are Android-capable (`Core` + `Avalonia`; WinForms excluded), the **SkiaSharp native-version pin** (`SkiaSharp.NativeAssets.Android` must match Avalonia's bundled `2.88.x`), **scoped-storage / SAF** ROM access (the desktop file dialogs collapse picks to a local path that Android `content://` URIs lack), **`config/` bundling** (ship as `AndroidAsset`, extract to `Context.FilesDir` on first run), the **multi-window → single-activity** navigation gap (the largest item), and the build prerequisites. An authored Avalonia.Android head **skeleton** lives at [`FEBuilderGBA.Android/`](../FEBuilderGBA.Android/README.md) — intentionally **not** in `FEBuilderGBA.sln` so it never affects the desktop CI build.
+The structural prerequisite **landed in [#1121](https://github.com/laqieer/FEBuilderGBA/issues/1121)**: `FEBuilderGBA.Avalonia` now **conditionally multi-targets** `net9.0;net9.0-android` — opt-in via the `EnableAndroidTarget` MSBuild property (default **OFF**, so the desktop `.sln` and all CI still see `net9.0` only). The Android head at [`FEBuilderGBA.Android/`](../FEBuilderGBA.Android/README.md) **builds a real APK** against the shared UI:
+
+```bash
+dotnet workload install android
+dotnet build FEBuilderGBA.Android/FEBuilderGBA.Android.csproj -c Release -p:EnableAndroidTarget=true
+```
+
+The `-p:EnableAndroidTarget=true` flag is required: the `ProjectReference`'s `AdditionalProperties` activates the android TFM for the build phase, but NuGet restore ignores per-reference `AdditionalProperties`, so the global property is needed to make restore cross-target the shared project too. The APK is **not yet device-validated**: booting it shows no editor until the single-view navigation rework lands. Desktop-only pieces (`Program.cs`, `Avalonia.Desktop`, `app.manifest`, the `GapSweep` Roslyn dev-tooling) are excluded on the android TFM.
+
+The full, evidence-backed feasibility assessment lives in **[docs/ANDROID.md](ANDROID.md)**: which layers are Android-capable (`Core` + `Avalonia`; WinForms excluded), the **SkiaSharp native-version pin** (`SkiaSharp.NativeAssets.Android` must match Avalonia's bundled `2.88.x`), **scoped-storage / SAF** ROM access (the desktop file dialogs collapse picks to a local path that Android `content://` URIs lack), **`config/` bundling** (ship as `AndroidAsset`, extract to `Context.FilesDir` on first run), the **multi-window → single-activity** navigation gap (the largest item), and the build prerequisites. The head is intentionally **not** in `FEBuilderGBA.sln` so it never affects the desktop CI build.
